Minor clean-ups
- Moves `CellCompleter` trait and object to separate file.
- Removes unused `Dep` class.
